Fix ladders snapping, fix 2 incorrect death animations
Partially fixed via Lua the ladders snapping which occured when Lara stopped going up or down in the middle of the animation.

Fixed the swandive roll animation being triggered instead of the death animation when landing during a somersault.

Fixed a rope animation triggering instead of the death animation, when Lara died during one of her stop-running animations.